Moved wrap.py into ciphers.modi
