+++ src/ciphers/modi/wrap.py
